reuse counter
