+++ src/utils/task_stats_alloc/Cargo.toml
@@ -15,10 +15,12 @@ tokio = { version = "0.2", package = "madsim-tokio", features = [
     "time",
     "signal",
 ] }
-#workspace-hack = { path = "../../workspace-hack" }
 
 [dev-dependencies]
 
 
 [target.'cfg(loom)'.dependencies]
 loom = {version = "0.5", features = ["futures", "checkpoint"]}
+
+[target.'cfg(not(loom))'.dependencies]
+workspace-hack = { path = "../../workspace-hack" }
